Removed the Group folder. Instead, because AffinityGroups and Riders are a many to many relationship, the Rider model was updated. Also removed the 'group' folder from the Django configuration in dataCollection/settings.py
